gall_add_transforms
Add a new key 'transform' to the datasource dictionary.  The value is
the name of a file which is evaluated by the collector.  The file must
contain a perl snippet that returns a function which takes a value as
input, performs a transformation on it and returns the result.

The function is called right after the collector has fetched the
datasource for a particular target and before it is written to the RRD
file.