gall_lookup_instance_mapping
This patch implements new semantics for the instance mapping feature.
If the new key "lookup" is set to true and not match key is present,
the instance is obtained directly from the value of the baseOID.

An example usage is for the CISCO-NTP-MIB, where cntpSysPeer holds the
association ID of the NTP system peer, which is used as an index into
the cntpPeersVarTable.